pyproject.toml
[project]
name = "AreaGrade"
version = "0.1.0"
description = "Web app to grade areas in Antwerp based on public transport accessibility"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "datasets>=3.5.1",
    "flask>=2.0.0",
    "folium>=0.12.0",
    "geopy>=2.2.0",
    "google-genai>=1.13.0",
    "google-generativeai>=0.8.5",
    "ipykernel>=6.29.5",
    "matplotlib>=3.10.1",
    "numpy>=1.24.0",
    "openai>=1.77.0",
    "pandas>=2.2.3",
    "python-dotenv>=1.1.0",
    "requests>=2.28.0",
]

[tool.rye.dependencies]
datasets = "*"
pandas = "*"
